Breaking: fix off() behavior, sync with socket-io-client
off() without an event name just unregisters all event handlers via
EventEmitter.

The comment implied offAny() would do this, but it's not the case:
that just removes the onAny() handlers.

If we really wanted to remove all event handlers for everything we
should be calling:
 - offAny()
 - offAnyOutgoing()
 - off()

But that would not match the wrapped behavior, so let's just keep it
simple.